Refactor: generic buttonSegment
